Move logFilter components & extend textField themed component
